+++ src/modules/reporting/components/common/disputing-markets.jsx
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import NullStateMessage from "modules/common/components/null-state-message/null-state-message";
 import DisputeMarketCard from "modules/reporting/components/dispute-market-card/dispute-market-card";
 import Paginator from "modules/common/components/paginator/paginator";
-import MarketsHeaderLabel from "modules/markets/components/markets-header-label/markets-header-label";
+import MarketsHeaderLabel from "modules/markets-list/components/markets-header-label/markets-header-label";
 import isEqual from "lodash/isEqual";
 
 export default class DisputingMarkets extends Component {
